+++ app/helpers/application_helper.rb
@@ -1,2 +1,23 @@
 module ApplicationHelper
+  def current_user_session
+      return @current_user_session if defined?(@current_user_session)
+      @current_user_session = UserSession.find
+    end
+
+    def current_user
+      return @current_user if defined?(@current_user)
+      @current_user = current_user_session && current_user_session.user
+    end
+
+    def require_login
+      unless current_user
+        redirect_to login_url
+        return false
+      end
+    end
+
+    def logged_in?
+      current_user_session != nil
+    end
+
 end
