+++ app/models/comment.rb
@@ -1,4 +1,5 @@
 class Comment < ActiveRecord::Base
   belongs_to :user
   belongs_to :essay
+  default_scope {order("updated_at desc")}
 end
