+++ app/models/tag.rb
@@ -1,0 +1,2 @@
+class Tag < ActiveRecord::Base
+end
