+++ db/migrate/20150818022048_create_tags.rb
@@ -1,0 +1,8 @@
+class CreateTags < ActiveRecord::Migration
+  def change
+    create_table :tags do |t|
+      t.string :name
+      t.timestamps null: false
+    end
+  end
+end
