+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20150817080209) do
+ActiveRecord::Schema.define(version: 20150818022048) do
 
   create_table "comments", force: :cascade do |t|
     t.integer  "essay_id"
@@ -54,6 +54,11 @@ ActiveRecord::Schema.define(version: 20150817080209) do
   create_table "posts", force: :cascade do |t|
     t.integer  "user_id"
     t.string   "content"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  create_table "tags", force: :cascade do |t|
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
